Multiplayer progress
It is still really bugged, as it is still creating too many other players and it leaves an after image of the other players where they were when the extra was created. It also doesn't seem to be giving rotational data properly. but it does however actively track player positional data properly.